refactor: moved encoding functions and packet interface
