+++ encoding.go
@@ -1,0 +1,136 @@
+package osc
+
+import (
+	"bufio"
+	"bytes"
+	"encoding/binary"
+	"fmt"
+	"strings"
+)
+
+// readBlob reads an OSC blob from the blob byte array. Padding bytes are
+// removed from the reader and not returned.
+func readBlob(reader *bufio.Reader) ([]byte, int, error) {
+	// First, get the length
+	var blobLen int32
+	if err := binary.Read(reader, binary.BigEndian, &blobLen); err != nil {
+		return nil, 0, err
+	}
+	n := 4 + int(blobLen)
+
+	if blobLen < 1 || blobLen > int32(reader.Buffered()) {
+		return nil, 0, fmt.Errorf("readBlob: invalid blob length %d", blobLen)
+	}
+
+	// Read the data
+	blob := make([]byte, blobLen)
+	if _, err := reader.Read(blob); err != nil {
+		return nil, 0, err
+	}
+
+	// Remove the padding bytes
+	numPadBytes := padBytesNeeded(int(blobLen))
+	if numPadBytes > 0 {
+		n += numPadBytes
+		dummy := make([]byte, numPadBytes)
+		if _, err := reader.Read(dummy); err != nil {
+			return nil, 0, err
+		}
+	}
+
+	return blob, n, nil
+}
+
+// writeBlob writes the data byte array as an OSC blob into buff. If the length
+// of data isn't 32-bit aligned, padding bytes will be added.
+func writeBlob(data []byte, buf *bytes.Buffer) (int, error) {
+	// Add the size of the blob
+	lenData := len(data)
+	err := binary.Write(buf, binary.BigEndian, int32(lenData))
+	if err != nil {
+		return 0, err
+	}
+
+	// Write the data
+	_, err = buf.Write(data)
+	if err != nil {
+		return 0, err
+	}
+
+	// Add padding bytes if necessary
+	numPadBytes := padBytesNeeded(lenData)
+	if numPadBytes > 0 {
+		padBytes := make([]byte, numPadBytes)
+		n, err := buf.Write(padBytes)
+		if err != nil {
+			return 0, err
+		}
+		numPadBytes = n
+	}
+
+	return 4 + lenData + numPadBytes, nil
+}
+
+// readPaddedString reads a padded string from the given reader. The padding
+// bytes are removed from the reader.
+func readPaddedString(reader *bufio.Reader) (string, int, error) {
+	// Read the string from the reader
+	str, err := reader.ReadString(0)
+	if err != nil {
+		return "", 0, err
+	}
+	lenStr := len(str)
+	n := lenStr
+
+	// Remove the padding bytes (leaving the null delimiter)
+	padLen := padBytesNeeded(lenStr)
+	if padLen > 0 {
+		n += padLen
+		padBytes := make([]byte, padLen)
+		if _, err = reader.Read(padBytes); err != nil {
+			return "", 0, err
+		}
+	}
+
+	// Strip off the string delimiter
+	return str[:lenStr-1], n, nil
+}
+
+// writePaddedString writes a string with padding bytes to the a buffer.
+// Returns, the number of written bytes and an error if any.
+func writePaddedString(str string, buf *bytes.Buffer) (int, error) {
+	// Truncate at the first null, just in case there is more than one present
+	nullIndex := strings.Index(str, "\x00")
+	if nullIndex > 0 {
+		str = str[:nullIndex]
+	}
+	// Write the string to the buffer
+	n, err := buf.WriteString(str)
+	if err != nil {
+		return 0, err
+	}
+
+	// Always write a null terminator, as we stripped it earlier if it existed
+	buf.WriteByte(0)
+	n++
+
+	// Calculate the padding bytes needed and create a buffer for the padding bytes
+	numPadBytes := padBytesNeeded(n)
+	if numPadBytes > 0 {
+		padBytes := make([]byte, numPadBytes)
+		// Add the padding bytes to the buffer
+		n, err := buf.Write(padBytes)
+		if err != nil {
+			return 0, err
+		}
+		numPadBytes = n
+	}
+
+	return n + numPadBytes, nil
+}
+
+// padBytesNeeded determines how many bytes are needed to fill up to the next 4
+// byte length.
+func padBytesNeeded(elementLen int) int {
+	return ((4 - (elementLen % 4)) % 4)
+}
