+++ osc.go
@@ -3,10 +3,6 @@
 package osc
 
 import (
-	"bufio"
-	"bytes"
-	"encoding/binary"
-	"fmt"
 	"regexp"
 	"strings"
 )
@@ -15,11 +11,6 @@ const (
 	secondsFrom1900To1970 = 2208988800
 	bundleTagString       = "#bundle"
 )
-
-// Packet is the interface for Message and Bundle.
-type Packet interface {
-	MarshalBinary() (data []byte, err error)
-}
 
 // Handler is an interface for message handlers. Every handler implementation
 // for an OSC message must implement this interface.
@@ -35,342 +26,6 @@ type HandlerFunc func(msg *Message)
 // Handler interface.
 func (f HandlerFunc) HandleMessage(msg *Message) {
 	f(msg)
-}
-
-////
-// StandardDispatcher
-////
-
-// receivePacket receives an OSC packet from the given reader.
-func readPacket(reader *bufio.Reader, start *int, end int) (Packet, error) {
-	// var buf []byte
-	buf, err := reader.Peek(1)
-	if err != nil {
-		return nil, err
-	}
-
-	switch buf[0] {
-	case '/':
-		return readMessage(reader, start)
-
-	case '#':
-		return readBundle(reader, start, end)
-	}
-
-	return nil, ERROR_INVALID_PACKET
-}
-
-// readBundle reads an Bundle from reader.
-func readBundle(reader *bufio.Reader, start *int, end int) (*Bundle, error) {
-	// Read the '#bundle' OSC string
-	startTag, n, err := readPaddedString(reader)
-	if err != nil {
-		return nil, err
-	}
-	*start += n
-
-	if startTag != bundleTagString {
-		return nil, fmt.Errorf("Invalid bundle start tag: %s", startTag)
-	}
-
-	// Read the timetag
-	var timeTag uint64
-	err = binary.Read(reader, binary.BigEndian, &timeTag)
-	if err != nil {
-		return nil, err
-	}
-
-	*start += 8
-
-	// Create a new bundle
-	bundle := NewBundle(timetagToTime(timeTag))
-
-	// Read until the end of the buffer
-	for *start < end {
-		// Read the size of the bundle element
-		var length int32
-
-		err = binary.Read(reader, binary.BigEndian, &length)
-		if err != nil {
-			return nil, err
-		}
-
-		*start += 4
-
-		p, err := readPacket(reader, start, end)
-		if err != nil {
-			return nil, err
-		}
-
-		err = bundle.Append(p)
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	return bundle, nil
-}
-
-// readMessage from `reader`.
-func readMessage(reader *bufio.Reader, start *int) (*Message, error) {
-	// First, read the OSC address
-	addr, n, err := readPaddedString(reader)
-	if err != nil {
-		return nil, err
-	}
-	*start += n
-
-	// Read all arguments
-	msg := NewMessage(addr)
-
-	err = readArguments(msg, reader, start)
-	if err != nil {
-		return nil, err
-	}
-
-	return msg, nil
-}
-
-// readArguments from `reader` and add them to the OSC message `msg`.
-func readArguments(msg *Message, reader *bufio.Reader, start *int) error {
-	// Read the type tag string
-	var n int
-	typetags, n, err := readPaddedString(reader)
-	if err != nil {
-		return err
-	}
-	*start += n
-
-	if len(typetags) == 0 {
-		return nil
-	}
-
-	// If the typetag doesn't start with ',', it's not valid
-	if typetags[0] != ',' {
-		return fmt.Errorf("unsupported type tag string %s", typetags)
-	}
-
-	// Remove ',' from the type tag
-	typetags = typetags[1:]
-
-	for _, c := range typetags {
-		switch c {
-		case 'i': // int32
-			var i int32
-
-			err = binary.Read(reader, binary.BigEndian, &i)
-			if err != nil {
-				return err
-			}
-
-			*start += 4
-			msg.Append(i)
-
-		case 'h': // int64
-			var i int64
-			err = binary.Read(reader, binary.BigEndian, &i)
-			if err != nil {
-				return err
-			}
-			*start += 8
-			msg.Append(i)
-
-		case 'f': // float32
-			var f float32
-			err = binary.Read(reader, binary.BigEndian, &f)
-			if err != nil {
-				return err
-			}
-			*start += 4
-			msg.Append(f)
-
-		case 'd': // float64/double
-			var d float64
-			err = binary.Read(reader, binary.BigEndian, &d)
-			if err != nil {
-				return err
-			}
-			*start += 8
-			msg.Append(d)
-
-		case 's': // string
-			// TODO: fix reading string value
-			var s string
-			s, _, err = readPaddedString(reader)
-			if err != nil {
-				return err
-			}
-			lenStr := len(s)
-			*start += lenStr + padBytesNeeded(lenStr)
-			msg.Append(s)
-
-		case 'b': // blob
-			var buf []byte
-			var n int
-			buf, n, err = readBlob(reader)
-			if err != nil {
-				return err
-			}
-			*start += n
-			msg.Append(buf)
-
-		case 't': // OSC time tag
-			var tt uint64
-
-			err = binary.Read(reader, binary.BigEndian, &tt)
-			if err != nil {
-				return nil
-			}
-
-			*start += 8
-			msg.Append(*NewTimetagFromTimetag(tt))
-
-		case 'N': // nil
-			msg.Append(nil)
-
-		case 'T': // true
-			msg.Append(true)
-
-		case 'F': // false
-			msg.Append(false)
-
-		default:
-			return fmt.Errorf("unsupported type tag: %c", c)
-		}
-	}
-
-	return nil
-}
-
-////
-// De/Encoding functions
-////
-
-// readBlob reads an OSC blob from the blob byte array. Padding bytes are
-// removed from the reader and not returned.
-func readBlob(reader *bufio.Reader) ([]byte, int, error) {
-	// First, get the length
-	var blobLen int32
-	if err := binary.Read(reader, binary.BigEndian, &blobLen); err != nil {
-		return nil, 0, err
-	}
-	n := 4 + int(blobLen)
-
-	if blobLen < 1 || blobLen > int32(reader.Buffered()) {
-		return nil, 0, fmt.Errorf("readBlob: invalid blob length %d", blobLen)
-	}
-
-	// Read the data
-	blob := make([]byte, blobLen)
-	if _, err := reader.Read(blob); err != nil {
-		return nil, 0, err
-	}
-
-	// Remove the padding bytes
-	numPadBytes := padBytesNeeded(int(blobLen))
-	if numPadBytes > 0 {
-		n += numPadBytes
-		dummy := make([]byte, numPadBytes)
-		if _, err := reader.Read(dummy); err != nil {
-			return nil, 0, err
-		}
-	}
-
-	return blob, n, nil
-}
-
-// writeBlob writes the data byte array as an OSC blob into buff. If the length
-// of data isn't 32-bit aligned, padding bytes will be added.
-func writeBlob(data []byte, buf *bytes.Buffer) (int, error) {
-	// Add the size of the blob
-	lenData := len(data)
-	err := binary.Write(buf, binary.BigEndian, int32(lenData))
-	if err != nil {
-		return 0, err
-	}
-
-	// Write the data
-	_, err = buf.Write(data)
-	if err != nil {
-		return 0, err
-	}
-
-	// Add padding bytes if necessary
-	numPadBytes := padBytesNeeded(lenData)
-	if numPadBytes > 0 {
-		padBytes := make([]byte, numPadBytes)
-		n, err := buf.Write(padBytes)
-		if err != nil {
-			return 0, err
-		}
-		numPadBytes = n
-	}
-
-	return 4 + lenData + numPadBytes, nil
-}
-
-// readPaddedString reads a padded string from the given reader. The padding
-// bytes are removed from the reader.
-func readPaddedString(reader *bufio.Reader) (string, int, error) {
-	// Read the string from the reader
-	str, err := reader.ReadString(0)
-	if err != nil {
-		return "", 0, err
-	}
-	lenStr := len(str)
-	n := lenStr
-
-	// Remove the padding bytes (leaving the null delimiter)
-	padLen := padBytesNeeded(lenStr)
-	if padLen > 0 {
-		n += padLen
-		padBytes := make([]byte, padLen)
-		if _, err = reader.Read(padBytes); err != nil {
-			return "", 0, err
-		}
-	}
-
-	// Strip off the string delimiter
-	return str[:lenStr-1], n, nil
-}
-
-// writePaddedString writes a string with padding bytes to the a buffer.
-// Returns, the number of written bytes and an error if any.
-func writePaddedString(str string, buf *bytes.Buffer) (int, error) {
-	// Truncate at the first null, just in case there is more than one present
-	nullIndex := strings.Index(str, "\x00")
-	if nullIndex > 0 {
-		str = str[:nullIndex]
-	}
-	// Write the string to the buffer
-	n, err := buf.WriteString(str)
-	if err != nil {
-		return 0, err
-	}
-
-	// Always write a null terminator, as we stripped it earlier if it existed
-	buf.WriteByte(0)
-	n++
-
-	// Calculate the padding bytes needed and create a buffer for the padding bytes
-	numPadBytes := padBytesNeeded(n)
-	if numPadBytes > 0 {
-		padBytes := make([]byte, numPadBytes)
-		// Add the padding bytes to the buffer
-		n, err := buf.Write(padBytes)
-		if err != nil {
-			return 0, err
-		}
-		numPadBytes = n
-	}
-
-	return n + numPadBytes, nil
-}
-
-// padBytesNeeded determines how many bytes are needed to fill up to the next 4
-// byte length.
-func padBytesNeeded(elementLen int) int {
-	return ((4 - (elementLen % 4)) % 4)
 }
 
 ////
